Refactor

+++ application/models/User_model.php
@@ -8,9 +8,9 @@ defined('BASEPATH') OR exit('No direct script access allowed');
 
 class User_model extends CI_Model{
 
-	public function __construct(){
+	public function __construct()
+	{
 		parent::__construct();
-		$this->load->database();
 	}
 
 
@@ -18,9 +18,15 @@ class User_model extends CI_Model{
 
 
 	/**
+	 * Have User
+	 *
 	 * Returns TRUE if there is a user with username $username in database
+	 *
+	 * @param $username
+	 * @return bool
 	 */
-	public function have_user($username){
+	public function have_user($username)
+	{
 		$query = $this->db->get_where('users', array('username'=>$username));
 		if ($query->num_rows() == 0)
 			return FALSE;
@@ -34,9 +40,15 @@ class User_model extends CI_Model{
 
 
 	/**
+	 * User ID to Username
+	 *
 	 * Converts user id to username (returns FALSE if user does not exist)
+	 *
+	 * @param $user_id
+	 * @return bool
 	 */
-	public function user_id_to_username($user_id){
+	public function user_id_to_username($user_id)
+	{
 		if( ! is_numeric($user_id))
 			return FALSE;
 		$query = $this->db->select('username')->get_where('users', array('id'=>$user_id));
@@ -50,9 +62,36 @@ class User_model extends CI_Model{
 
 
 	/**
-	 * Returns TRUE if there is a user (except $username) with email $email in database
+	 * Username to User ID
+	 *
+	 * Converts username to user id (returns FALSE if user does not exist)
+	 *
+	 * @param $username
+	 * @return bool
 	 */
-	public function have_email($email, $username = FALSE){
+	public function username_to_user_id($username)
+	{
+		$query = $this->db->select('id')->get_where('users', array('username'=>$username));
+		if ($query->num_rows() == 0)
+			return FALSE;
+		return $query->row()->id;
+	}
+
+
+	// ------------------------------------------------------------------------
+
+
+	/**
+	 * Have Email
+	 *
+	 * Returns TRUE if a user (except $username) with given email exists
+	 *
+	 * @param $email
+	 * @param bool $username
+	 * @return bool
+	 */
+	public function have_email($email, $username = FALSE)
+	{
 		$query = $this->db->get_where('users', array('email'=>$email));
 		if ($query->num_rows() >= 1){
 			if($username !== FALSE && $query->row()->username == $username)
@@ -68,9 +107,18 @@ class User_model extends CI_Model{
 
 
 	/**
-	 * Add a new user to database
+	 * Add User
+	 *
+	 * Adds a single user
+	 *
+	 * @param $username
+	 * @param $email
+	 * @param $password
+	 * @param $role
+	 * @return bool|string
 	 */
-	public function add_user($username, $email, $password, $role){
+	public function add_user($username, $email, $password, $role)
+	{
 		if (strlen($username) < 3 OR strlen($username) > 20 OR strlen($password) < 6 OR strlen($password) > 200)
 			return 'Username or password length error.';
 		if ($this->have_user($username))
@@ -98,17 +146,25 @@ class User_model extends CI_Model{
 
 
 	/**
-	 * Add multiple users
+	 * Add Users
+	 *
+	 * Adds multiple users
+	 *
+	 * @param $text
+	 * @param $send_mail
+	 * @param $delay
+	 * @return array
 	 */
-	public function add_users($text, $send_mail, $delay){
+	public function add_users($text, $send_mail, $delay)
+	{
 
 		$lines = preg_split('/\r?\n|\n?\r/', $text);
 		$users_ok = array();
 		$users_error = array();
 
 		// loop over lines of $text :
-		foreach ($lines as $line){
-
+		foreach ($lines as $line)
+		{
 			$line = trim($line);
 
 			if (strlen($line) == 0 OR $line[0] == '#')
@@ -118,7 +174,9 @@ class User_model extends CI_Model{
 			if (count($parts) != 4)
 				continue; //ignore lines that not contain 4 parts
 
-			if (strtolower(substr($parts[2], 0, 6)) == 'random'){ // generate random password
+			if (strtolower(substr($parts[2], 0, 6)) == 'random')
+			{
+				// generate random password
 				$len = trim(substr($parts[2], 6), '[]');
 				if (is_numeric($len)){
 					$this->load->helper('string');
@@ -135,7 +193,9 @@ class User_model extends CI_Model{
 
 		} // end of loop
 
-		if ($send_mail){ // sending usernames and passwords by email
+		if ($send_mail)
+		{
+			// sending usernames and passwords by email
 			$this->load->library('email');
 			$config = array(
 				'mailtype'  => 'html',
@@ -157,7 +217,8 @@ class User_model extends CI_Model{
 			$this->email->set_newline("\r\n");
 			$count_users = count($users_ok);
 			$counter = 0;
-			foreach ($users_ok as $user){
+			foreach ($users_ok as $user)
+			{
 				$counter++;
 				$this->email->from($this->settings_model->get_setting('mail_from'), $this->settings_model->get_setting('mail_from_name'));
 				$this->email->to($user[1]);
@@ -184,18 +245,36 @@ class User_model extends CI_Model{
 
 
 	/**
-	 * Delete a user from database
+	 * Delete User
+	 *
+	 * Deletes a user with given user id
+	 * Returns TRUE (success) or FALSE (failure)
+	 *
+	 * @param $user_id
+	 * @return bool
 	 */
-	public function delete_user($username){
-		// Delete from database
-		$this->db->delete('users', array('username'=>$username));
+	public function delete_user($user_id)
+	{
+		$this->db->trans_start();
+
+		$username = $this->user_id_to_username($user_id);
+		if ($username === FALSE)
+			return FALSE;
+		$this->db->delete('users', array('id'=>$user_id));
 		$this->db->delete('final_submissions', array('username' => $username));
 		$this->db->delete('all_submissions', array('username' => $username));
-		// Delete submitted files
-		shell_exec("cd {$this->settings_model->get_setting('assignments_root')}; rm -r */*/{$username};");
 		// each time we delete a user, we should update all scoreboards
 		$this->load->model('scoreboard_model');
 		$this->scoreboard_model->update_scoreboards();
+
+		$this->db->trans_complete();
+
+		if ($this->db->trans_status()) {
+			// Delete submitted files
+			shell_exec("cd {$this->settings_model->get_setting('assignments_root')}; rm -r */*/{$username};");
+			return TRUE; //success
+		}
+		return FALSE; // failure
 	}
 
 
@@ -203,9 +282,54 @@ class User_model extends CI_Model{
 
 
 	/**
-	 * Function used for validating user login
+	 * Delete Submissions
+	 *
+	 * Deletes all submissions of user with given user id
+	 * Returns TRUE (success) or FALSE (failure)
+	 *
+	 * @param $user_id
+	 * @return bool
 	 */
-	public function validate_user($username, $password){
+	public function delete_submissions($user_id)
+	{
+		$this->db->trans_start();
+
+		$username = $this->user_id_to_username($user_id);
+		if ($username === FALSE)
+			return FALSE;
+		// delete all submissions from database
+		$this->db->delete('final_submissions', array('username'=>$username));
+		$this->db->delete('all_submissions', array('username'=>$username));
+		// each time we delete a user's submissions, we should update all scoreboards
+		$this->load->model('scoreboard_model');
+		$this->scoreboard_model->update_scoreboards();
+
+		$this->db->trans_complete();
+
+		if ($this->db->trans_status()) {
+			// delete all submitted files
+			shell_exec("cd {$this->settings_model->get_setting('assignments_root')}; rm -r */*/{$username};");
+			return TRUE; // success
+		}
+
+		return FALSE; // failure
+	}
+
+
+	// ------------------------------------------------------------------------
+
+
+	/**
+	 * Validate User
+	 *
+	 * Returns TRUE if given username and password is valid for login
+	 *
+	 * @param $username
+	 * @param $password
+	 * @return bool
+	 */
+	public function validate_user($username, $password)
+	{
 		$this->load->library('password_hash', array(8, FALSE));
 		$query = $this->db->get_where('users', array('username' => $username));
 		if ($query->num_rows() != 1)
@@ -221,7 +345,15 @@ class User_model extends CI_Model{
 	// ------------------------------------------------------------------------
 
 
-	public function update_login_time($username) {
+	/**
+	 * Update Login Time
+	 *
+	 * Updates First Login Time and Last Login Time for given username
+	 *
+	 * @param $username
+	 */
+	public function update_login_time($username)
+	{
 		$now = date('Y-m-d H:i:s', shj_now());
 
 		$first_login = $this->db->select('first_login_time')->get_where('users', array('username'=>$username))->row()->first_login_time;
@@ -231,13 +363,19 @@ class User_model extends CI_Model{
 		$this->db->where('username', $username)->update('users', array('last_login_time'=>$now));
 	}
 
+
 	// ------------------------------------------------------------------------
 
 
 	/**
-	 * Returns selected assignment by user $username
+	 * Selected Assignment
+	 *
+	 * Returns selected assignment by given username
+	 * @param $username
+	 * @return mixed
 	 */
-	public function selected_assignment($username){
+	public function selected_assignment($username)
+	{
 		$query = $this->db->select('selected_assignment')->get_where('users', array('username'=>$username));
 		if ($query->num_rows() != 1){//logout
 			$this->session->sess_destroy();
@@ -251,9 +389,15 @@ class User_model extends CI_Model{
 
 
 	/**
+	 * Select Assignment
+	 *
 	 * Sets selected assignment for $username
+	 *
+	 * @param $username
+	 * @param $assignment_id
 	 */
-	public function select_assignment($username, $assignment_id){
+	public function select_assignment($username, $assignment_id)
+	{
 		$this->db->where('username', $username)->update('users', array('selected_assignment'=>$assignment_id));
 	}
 
@@ -262,9 +406,15 @@ class User_model extends CI_Model{
 
 
 	/**
+	 * Get Display Name
+	 *
 	 * Returns name of the user with given username
+	 *
+	 * @param $username
+	 * @return bool
 	 */
-	public function get_display_name($username){
+	public function get_display_name($username)
+	{
 		$query = $this->db->select('display_name')->get_where('users', array('username'=>$username));
 		if ($query->num_rows() != 1)
 			return FALSE;
@@ -276,13 +426,19 @@ class User_model extends CI_Model{
 
 
 	/**
+	 * Get User Level
+	 *
 	 * Returns permission level of given user
 	 * admin            -> 3
 	 * head_instructor  -> 2
 	 * instructor       -> 1
 	 * student          -> 0
+	 *
+	 * @param $username
+	 * @return int
 	 */
-	public function get_user_level($username){
+	public function get_user_level($username)
+	{
 		$role = $this->db->select('role')->get_where('users', array('username'=>$username))->row()->role;
 		switch ($role){
 			case 'admin': return 3;
@@ -297,9 +453,15 @@ class User_model extends CI_Model{
 
 
 	/**
-	 * Update user profile
+	 * Update Profile
+	 *
+	 * Updates User Profile (Name, Email, Password, Role)
+	 *
+	 * @param $user_id
+	 * @return bool
 	 */
-	public function update_profile($user_id){
+	public function update_profile($user_id)
+	{
 		$query = $this->db->get_where('users', array('id'=>$user_id));
 		if ($query->num_rows() != 1)
 			return FALSE;
@@ -330,10 +492,15 @@ class User_model extends CI_Model{
 
 
 	/**
-	 * Generate a password reset key and
-	 * send an email containing the link for resetting password (in case of password lost)
+	 * Send Password Reset Mail
+	 *
+	 * Generates a password reset key and sends an email containing the link
+	 * for resetting password (in case of password lost)
+	 *
+	 * @param $email
 	 */
-	public function send_passchange_mail($email){
+	public function send_password_reset_mail($email)
+	{
 		// exit if $email is invalid:
 		if ( ! $this->have_email($email) )
 			return;
@@ -382,9 +549,16 @@ class User_model extends CI_Model{
 
 
 	/**
+	 * Password Reset Key Is Valid
+	 *
 	 * Returns TRUE if the given password reset key is valid
+	 * And returns an error message if key is invalid
+	 *
+	 * @param $passchange_key
+	 * @return bool|string
 	 */
-	public function have_passchange($passchange_key){
+	public function passchange_is_valid($passchange_key)
+	{
 		$query = $this->db->select('passchange_time')->get_where('users', array('passchange_key'=>$passchange_key));
 		if ($query->num_rows() != 1)
 			return 'Invalid password reset link.';
@@ -400,9 +574,16 @@ class User_model extends CI_Model{
 
 
 	/**
+	 * Reset Password
+	 *
 	 * Resets password for given password reset key (in case of lost password)
+	 *
+	 * @param $passchange_key
+	 * @param $newpassword
+	 * @return bool
 	 */
-	public function reset_password($passchange_key, $newpassword){
+	public function reset_password($passchange_key, $newpassword)
+	{
 		$query = $this->db->get_where('users', array('passchange_key'=>$passchange_key));
 		if ($query->num_rows() != 1)
 			return FALSE; //failure
@@ -416,9 +597,14 @@ class User_model extends CI_Model{
 
 
 	/**
-	 * Get All Users Table (for users page)
+	 * Get All Users
+	 *
+	 * Returns an array of all users (for Users page)
+	 *
+	 * @return mixed
 	 */
-	public function get_all_users(){
+	public function get_all_users()
+	{
 		return $this->db->order_by('role', 'asc')->get('users')->result_array();
 	}
 
@@ -427,9 +613,35 @@ class User_model extends CI_Model{
 
 
 	/**
-	 * Saves positions of dashboard widgets in database
+	 * Get User
+	 *
+	 * Returns database row for given user id
+	 *
+	 * @param $user_id
+	 * @return bool
 	 */
-	public function save_widget_positions($username, $positions){
+	public function get_user($user_id)
+	{
+		$query = $this->db->get_where('users', array('id'=>$user_id));
+		if ($query->num_rows() != 1)
+			return FALSE;
+		return $query->row();
+	}
+
+
+	// ------------------------------------------------------------------------
+
+
+	/**
+	 * Save Widget Positions
+	 *
+	 * Updates position of dashboard widgets in database
+	 *
+	 * @param $username
+	 * @param $positions
+	 */
+	public function save_widget_positions($username, $positions)
+	{
 		$this->db->where('username', $username)->update('users', array('dashboard_widget_positions'=>$positions));
 	}
 
@@ -438,10 +650,22 @@ class User_model extends CI_Model{
 
 
 	/**
+	 * Get Widget Positions
+	 *
 	 * Returns positions of dashboard widgets from database
+	 *
+	 * @param $username
+	 * @return mixed
 	 */
-	public function get_widget_positions($username){
-		return json_decode($this->db->select('dashboard_widget_positions')->get_where('users', array('username'=>$username))->row()->dashboard_widget_positions, TRUE);
+	public function get_widget_positions($username)
+	{
+		return json_decode(
+			$this->db->select('dashboard_widget_positions')
+			->get_where('users', array('username'=>$username))
+			->row()
+			->dashboard_widget_positions,
+			TRUE
+		);
 	}
 
 }
